LPD-49619 Merge SF checks
