+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/check/JSONCatalogFileCheck.java
@@ -6,7 +6,6 @@
 package com.liferay.source.formatter.check;
 
 import com.liferay.portal.json.JSONArrayImpl;
-import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
@@ -17,7 +16,7 @@ import java.util.Comparator;
 /**
  * @author Alan Huang
  */
-public class JSONProductsSpecificationsFileCheck extends BaseFileCheck {
+public class JSONCatalogFileCheck extends BaseFileCheck {
 
 	@Override
 	public boolean isLiferaySourceCheck() {
@@ -29,15 +28,22 @@ public class JSONProductsSpecificationsFileCheck extends BaseFileCheck {
 			String fileName, String absolutePath, String content)
 		throws JSONException {
 
-		if (!absolutePath.endsWith(".products.specifications.json")) {
-			return content;
+		if (absolutePath.endsWith(".options.json") ||
+			absolutePath.endsWith(".products.json")) {
+
+			return JSONUtil.toString(
+				JsonSourceUtil.sortJSONArray(
+					new JSONArrayImpl(content),
+					new ExternalReferenceCodeComparator()));
+		}
+		else if (absolutePath.endsWith(".products.specifications.json")) {
+			return JSONUtil.toString(
+				JsonSourceUtil.sortJSONArray(
+					new JSONArrayImpl(content),
+					new CPDefinitionExternalReferenceCodeComparator()));
 		}
 
-		JSONArray jsonArray = new JSONArrayImpl(content);
-
-		return JSONUtil.toString(
-			JsonSourceUtil.sortJSONArray(
-				jsonArray, new CPDefinitionExternalReferenceCodeComparator()));
+		return content;
 	}
 
 	private class CPDefinitionExternalReferenceCodeComparator
@@ -66,6 +72,26 @@ public class JSONProductsSpecificationsFileCheck extends BaseFileCheck {
 			String key2 = jsonObject2.getString("key");
 
 			return key1.compareTo(key2);
+		}
+
+	}
+
+	private class ExternalReferenceCodeComparator
+		implements Comparator<Object> {
+
+		@Override
+		public int compare(Object object1, Object object2) {
+			JSONObject jsonObject1 = (JSONObject)object1;
+
+			String externalReferenceCode1 = jsonObject1.getString(
+				"externalReferenceCode");
+
+			JSONObject jsonObject2 = (JSONObject)object2;
+
+			String externalReferenceCode2 = jsonObject2.getString(
+				"externalReferenceCode");
+
+			return externalReferenceCode1.compareTo(externalReferenceCode2);
 		}
 
 	}
